[REFACTOR] #17 DataSource Interface로 추상화
