+++ app/src/main/java/com/example/android_accountbook_13/data/local/datasource/LocalDataSourceImpl.kt
@@ -1,31 +1,32 @@
-package com.example.android_accountbook_13.data.local
+package com.example.android_accountbook_13.data.local.datasource
 
 import android.database.Cursor
+import com.example.android_accountbook_13.data.local.SQLiteOpenHelper
 
-class LocalDataSource (
+class LocalDataSourceImpl(
     private val db: SQLiteOpenHelper
-) {
-    fun getHistory(month: Int): Cursor =
+): LocalDataSource {
+    override fun getHistory(month: Int): Cursor =
         db.readableDatabase.rawQuery(
             "SELECT * FROM payment WHERE month like ${month};",null
         )
 
-    fun getAllCategory(): Cursor =
+    override fun getAllCategory(): Cursor =
         db.readableDatabase.rawQuery(
             "SELECT * FROM category;",null
         )
 
-    fun getCategory(id: Int): Cursor =
+    override fun getCategory(id: Int): Cursor =
         db.readableDatabase.rawQuery(
             "SELECT * FROM category WHERE id like ${id};",null
         )
 
-    fun getAllPaymentMethod(): Cursor =
+    override fun getAllPaymentMethod(): Cursor =
         db.readableDatabase.rawQuery(
             "SELECT * FROM payment_method;",null
         )
 
-    fun getPaymentMethod(id: Int): Cursor =
+    override fun getPaymentMethod(id: Int): Cursor =
         db.readableDatabase.rawQuery(
             "SELECT * FROM payment_method WHERE id like ${id};",null
         )
