+++ app/src/main/java/com/example/android_accountbook_13/di/DataSourceModule.kt
@@ -1,7 +1,7 @@
 package com.example.android_accountbook_13.di
 
-import com.example.android_accountbook_13.data.local.repository.LocalRepositoryImpl
 import com.example.android_accountbook_13.data.local.repository.LocalRepository
+import com.example.android_accountbook_13.data.local.repository.LocalRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -10,9 +10,9 @@ import dagger.hilt.components.SingletonComponent
 @InstallIn(SingletonComponent::class)
 @Module
 abstract
-class RepositoryModule {
+class DataSourceModule {
     @Binds
-    abstract fun provideLocalRepository(
-        repositoryImpl: LocalRepositoryImpl
+    abstract fun provideLocalDataSource(
+        localRepositoryImpl: LocalRepositoryImpl
     ): LocalRepository
 }
